+++ src/components/metadata/Metadata.tsx
@@ -12,11 +12,8 @@
 
 import ArrowDownBlue from '@manuscripts/assets/react/ArrowDownBlue'
 import {
-  Affiliation,
   ContainerInvitation,
   Contributor,
-  ContributorRole,
-  ObjectTypes,
   UserProfile,
 } from '@manuscripts/json-schema'
 import {
@@ -32,7 +29,7 @@ import { TitleEditorView } from '@manuscripts/title-editor'
 import React, { useCallback } from 'react'
 import styled from 'styled-components'
 
-import { buildAuthorsAndAffiliations } from '../../lib/authors'
+import { getMetaData } from '../../lib/authors'
 import { useStore } from '../../store'
 import { AddAuthorsModalContainer } from './AddAuthorsModalContainer'
 import { InvitationValues } from './AuthorInvitationForm'
@@ -155,29 +152,10 @@ export const Metadata: React.FunctionComponent<Props> = (props) => {
       trackModelMap: store.trackModelMap,
     }
   })
-  console.log('+++++++++++++++++++++')
-  console.log(trackModelMap)
 
-  const affiliationAndContributors: (Contributor | Affiliation)[] = []
-  const contributorRoles: ContributorRole[] = []
-  // KOD identicky s buildData.ts r.236 => zavolat jako funkci s parametrem kam predam trackModelMap
-  if (trackModelMap) {
-    for (const model of trackModelMap.values()) {
-      if (
-        model.objectType === ObjectTypes.Affiliation ||
-        model.objectType === ObjectTypes.Contributor
-      ) {
-        affiliationAndContributors.push(model as Affiliation) // or Contributor
-      }
-      if (model.objectType === ObjectTypes.ContributorRole) {
-        contributorRoles.push(model as ContributorRole)
-      }
-    }
-  }
-
-  const authorsAndAffiliations = buildAuthorsAndAffiliations(
-    affiliationAndContributors
-  )
+  const metaData = getMetaData(trackModelMap)
+  const authorsAndAffiliations = metaData?.authorsAndAffiliations
+  const contributorRoles = metaData?.contributorRoles || []
 
   const can = usePermissions()
