+++ src/hooks/use-tracked-model-management.ts
@@ -41,8 +41,6 @@ const useTrackedModelManagement = (
   finalModelMap: Map<string, Model>
 ) => {
   const modelMap = useMemo(() => {
-    console.log('use-tracked-model-management.ts:')
-    console.log(doc)
     const docJSONed = doc.toJSON()
     const docClean = adaptTrackedData(docJSONed)
     const modelsFromPM = encode(schema.nodeFromJSON(docClean))
@@ -189,7 +187,6 @@ const useTrackedModelManagement = (
         let foundInDoc = false
 
         let dataTrackedId = ''
-        console.log('model._id: ' + model._id)
         if (model._id?.includes(trackedJoint)) {
           // when encoding we modify ids of track changes artefacts to avoid duplicate ids in the modelMap
           // when saving back we need to convert those ids back and also apply the updates on the right nodes
